requirements.txt
# Insight Engine Python Dependencies

supabase>=2.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
openai